requirements.txt
langgraph
langchain_core
google-generativeai
python-dotenv
supabase
watchdog
tenacity
fastapi
uvicorn